backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Moi10legend",email = "moisesflopesf@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.126.0,<0.127.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "oracledb (>=3.4.1,<4.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "pytz (>=2025.2,<2026.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
